Revert "Ios  home  module"
