--- UserItem+CoreDataClass.swift
@@ -1,14 +1,0 @@
-//
-//  UserItem+CoreDataClass.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 25/06/22.
-//
-//
-
-import Foundation
-import CoreData
-
-public class UserItem: NSManagedObject {
-
-}
