--- UsersPublications/Common/EndpointList.swift
@@ -1,15 +1,0 @@
-//
-//  EndpointList.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 26/06/22.
-//
-
-import Foundation
-
-class EndpointList {
-    
-    // users
-    
-    static let GET_USERS_DATA = "https://jsonplaceholder.typicode.com/users"
-}
