--- UsersPublications/Common/Utils/Colors.swift
@@ -1,13 +1,0 @@
-//
-//  Colors.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 25/06/22.
-//
-
-import Foundation
-import UIKit
-
-struct Colors {
-    static let darkGreen = UIColor(named: "darkGreen", in: UsersPublicationsAPI.bundle, compatibleWith: nil)
-}
