--- UsersPublications/Modules/Configuration/UsersPublicationsAPI.swift
@@ -1,12 +1,0 @@
-//
-//  UsersPublicationsAPI.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 24/06/22.
-//
-
-import Foundation
-
-class UsersPublicationsAPI {
-    static let bundle = Bundle(for: UsersPublicationsAPI.self)
-}
