--- UsersPublications/Modules/Home/Data/HomeDataSource.swift
@@ -1,23 +1,0 @@
-//
-//  HomeDataSource.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 24/06/22.
-//
-
-import Foundation
-
-protocol IHomeDataSource {
-    func requestUsersData(endpoint: String, _ completion: @escaping (_ data: [[String: Any]]) -> Void)
-}
-
-class HomeDataSource: IHomeDataSource {
-    
-    private let requestManager = RequestManager()
-    
-    func requestUsersData(endpoint: String, _ completion: @escaping (_ data: [[String: Any]]) -> Void) {
-        requestManager.get(endpoint, parameters: nil) { data in
-            completion(data)
-        }
-    }
-}
