--- UsersPublications/Modules/Home/HomeRouter.swift
@@ -1,8 +1,0 @@
-//
-//  HomeRouter.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 24/06/22.
-//
-
-import Foundation
