--- UsersPublications/Modules/Home/Presentation/Views/DefaultCell.swift
@@ -1,16 +1,0 @@
-//
-//  DefaultCell.swift
-//  UsersPublications
-//
-//  Created by David Andres Mejia Lopez on 25/06/22.
-//
-
-import UIKit
-
-class DefaultCell: UITableViewCell {
-    static let id = "DefaultCell"
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-    }
-}
