+++ UsersPublications/Modules/Home/ViewController.swift
@@ -1,0 +1,17 @@
+//
+//  ViewController.swift
+//  UsersPublications
+//
+//  Created by David Andres Mejia Lopez on 24/06/22.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+}
+
